DAOS-10477 dfs: add new APIs to retrieve pool and container handles (#8868)
If a user connect to a DFS container with dfs_connect(), they don't have
access to the pool and container handles that are opened by DFS internally.
Users sometimes need to add attributes or query the pool and container info,
so provide a mechanism to return those internal pool and container handles
with proper refcounting, requiring users to put the handles when done before
umount of dfs.

Signed-off-by: Mohamad Chaarawi <mohamad.chaarawi@intel.com>
# ------------------------ >8 ------------------------
Skip-func-hw-test: true
Skip-func-test: true
Quick-Functional: true
Test-tag: dfuse