DAOS-10455 test: Enable enable_vmd: true whenever VMD devices exist. (#8834) (#8961)
Set enable_vmd: true in the daos_server yaml if there are VMD devices on
the server hosts regardless of whether or not they are being utilized in
the server config file.  This avoids errors with the server start-up
NVMe scan.  Also increased the pool/autotest.py timeout to fix test failure.

Signed-off-by: Phillip Henderson <phillip.henderson@intel.com>