DAOS-10465 vos: Avoid using aggregation optimization on v2.0 pool (#8953)
Explicitly enable the optimization when the GLOBAL_VERSION is >= 1.
DAOS 2.0 has no such property so the global version will be 0. This
should allow us to use a 2.0 pool even if it's been touched by 2.2

For pools that are created with 2.2, we set an new durable format version
such that this feature is always enabled. This maintains compatibility with
the old durable format up until a pool upgrade.

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>