DAOS-10138 pool: Try to restore some PS redundancy
Currently, a pool service (PS) only _replaces_ DOWN replicas. If there
are not enough engines in the pool, the PS runs at a degraded
redundancy. When the number of available engines in the pool increases,
the PS does not attempt to restore its redundancy. This patch slightly
improves this policy to try to restore _some_ redundancy.

Since we don't have a good way to get the original redundancy factor
(RF) of the PS, nor can we compute an ideal RF, we default to two (i.e.,
five replicas). If there are not enough engines in the pool, we reduce
the RF objective to one (i.e., three replicas), or zero (i.e., one
replica), if even one does not work.

We limit the scope of this patch to the policy, not the mechansim.

Signed-off-by: Li Wei <wei.g.li@intel.com>