DAOS-10615 vos: Improvements to Standalone VOS
If standalone VOS is using sys_db then it shouldn't delete the file
before creating or on exiting. Added function parameters so that
can indicate if using sys_db and also which target to use.
Previously for all standalone VOS instances always used -1 tgt id.
The daos debug tool will utilize the tgt id and will use the
sys_db smd file.

For initializing bio context, a target id of -1 was used to
indicate if it should self poll. However, to support the daos
debug tool, the actual target id is needed along with self polling.

Signed-off-by: Ryon Jensen <ryon.jensen@intel.com>