DAOS-10354 dfuse: Check thread id in interception library destructor.
Only run the destructor if run in the same thread id as the constructor.
some programs, at least sh, run lightweight threads so can end up
calling the destructor more times than the constructor, so to avoid
deadlocks save the tid in the constructor and only run the destructor
if it matches.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>