DAOS-623 build: Turn back on weak deps for Rocky installs.
Previously we had set the install_weak_deps option to false
in order to reduce CI image sizes however this got disable
a few weeks ago and re-enabled on the 19th, during the
time it was disabled however an upstream change means
the build no longer completes with this option enabled.

Disable the option again until we can work out which
weakly required packages are needed for the build.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>