DCO-10410 build: Improve handling of epel-release package.
The epel-release package exists in core CentOS, but also in
epel itself, so in order to get it up-to-date you need on a
new system you need to install the pacakge, then upgrade it
to get the newer epel version.  Have the dockerfiles action
this second install directly after the first.

Remove the makecache from REPOs sections of the dockerfiles
This ties in with other steps that delete any temporary and
cache files between RUN commands which causes the resulting
images to be smaller and avoids a extra download during the
build.

Together these avoid a bug we were seeing in docker/dnf for
cases where REPO was used, a combination of the cache files
and dnf updating epel-release meant that dnf was re-loading
the packages metatdata and causing a problem with overlayfs
where opening the same file RO and then RW would cause both
file descriptors to give differing version of the contents.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>