DAOS-10255 test: adjust minimum containers expectation again
Significantly reduce the test's expectation of the minimum number of
containers that can be created in a pool, since frequent changes to
metadata layouts have an impact that triggers test failures. And since
the tests are currently designed to fill the metadata  on purpose,
rather than expect a specific quantity.

Quick-Functional: true
Test-tag: metadata_fillup metadata_der_nospace pool_target_query pool_create_all_one_hw

Signed-off-by: Kenneth Cain <kenneth.c.cain@intel.com>

Required-githooks: true
Signed-off-by: Kenneth Cain <kenneth.c.cain@intel.com>