DAOS-9167 tools,control: add dmg pool query-targets (#9110)
Before this change libdaos API function daos_pool_query_target()
and the daos utility, "pool query-targets" command was provided to
inspect a given pool engine's targets (to see up/down state, storage
capacity/usage, etc.).

With this change, dmg pool query-targets is added to allow an admin to
inspect a given pool engine's targets.

Examples of the essential parts of the dmg command line:
  * dmg pool query-targets mypool --rank=6 --target-idx=3
  * dmg pool query-targets mypool --rank=0 --target-idx=0,1,5,7
  * dmg pool query-targets mypool -j --rank=0 --target-idx=0,1,5,7

A mgmt daos_server gRPC handler is implemented, and a corresponding
engine drpc handler that interacts with the pool service. The control
library adds a PoolQueryTarget method to initiate the gRPC from the
dmg utility side, and convert the protobuf reply to PoolQueryTargetResp
form for printing. And the new dmg command implementation is added.

Signed-off-by: Kenneth Cain <kenneth.c.cain@intel.com>