DAOS-7133 tse: refine task re-init checking
Original code using dtp_completing flag for task re-init check -
1. first set dtp_completing == 1 and call task completion CB,
2. inside task completion CB if it calls re-init then will zero the flag,
3. check dtp_completing flag if it is zero then the task was re-inited

But a problem is once the task re-inited in step2, it possibly get
executing and complete (which will set the dtp_completing flag)
immediately even before step 3.
Previously we added a lock dsp_comp_lock to avoid that but did not
resolve some cases' problem, and the lock is bad as it actually
serialized all tasks' completion callbacks' executing.

This patch refines that by adding an ATOMIC generation to the task,
and removes the dsp_comp_lock.

Signed-off-by: Xuezhao Liu <xuezhao.liu@intel.com>