DAOS-10621 test: Fix weekly-testing timed builds
Moving to using CHANGE_TARGET resolves the issue of running with the
correct daos RPMs in a weekly-testing* branch PR, but this variable is
not set during timed weekly-testing* builds.  Use the BRANCH_NAME
variable if the CHANGE_TARGET variable is not set to allow
weekly-testing* timed builds to install the correct doas RPMs.

Skip-unit-tests: true
Test-tag: always_passes

Signed-off-by: Phillip Henderson <phillip.henderson@intel.com>