DAOS-10643 control: Debounce SWIM Rank Dead events (#9236)
When a rank is seen as dead by multiple other ranks, the
number of events raised may be as large as N-1, where N
is the size of the cluster. As we only care about the first
event for the purpose of updating the system membership,
we can ignore subsequent events for the same rank:incarnation
combination.

Quick-build: true
Skip-test: true

Signed-off-by: Michael MacDonald <mjmac.macdonald@intel.com>