DAOS-10448 test: reduce SCM size in create_all_vm/other tests
Before this change, in virtual machine (VM)-based testing of the
pool create_all tests, intermittent failures were seen due to
slow/laggy performance on the VM nodes. This resulted in RPC timeouts
in daos_engine during pool create/destroy operations. This is not a
a new phenomenon. Previously it has been seen with other VM-based tests,
requiring a smaller scm_size setting to alleviate memory pressure.

With this change, the scm_size parameter is reduced in this test
and in the following tests (and addressing the associated tickets):
- container/query_properties (DAOS-10487)
- telemetry/dkey_akey_enum_punch (DAOS-10487)
- control/dmg_telemetry_basic (DAOS-10528)

Longer term for create_all_vm, it may be better to eliiminate it,
relying instead on the nearly-identical pool/create_all_hw.py tests.

Quick-Functional: true
Test-tag: pool_create_all,vm test_with_telemetry_basic container,query_properties dkey_akey_enum_punch tgt_dkey_akey_punch pool_tgt_dkey_akey_punch
Test-repeat: 10

Signed-off-by: Kenneth Cain <kenneth.c.cain@intel.com>

Required-githooks: true