DAOS-10448 test: reduce SCM size in create_all_vm/other tests (#9291) (#9300)
Cherry-pick of PR #9291 master commit c03a05623 to release/2.2 branch.

Before this change, in virtual machine (VM)-based testing of the
pool create_all tests, intermittent failures were seen due to
slow/laggy performance on the VM nodes. This resulted in RPC timeouts
in daos_engine during pool create/destroy operations. This is not a
a new phenomenon. Previously it has been seen with other VM-based tests,
requiring a smaller scm_size setting to alleviate memory pressure.

With this change, the scm_size parameter is reduced in this test
and in the following tests (and addressing the associated tickets):
- container/query_properties (DAOS-10487)
- telemetry/dkey_akey_enum_punch (DAOS-10487)
- control/dmg_telemetry_basic (DAOS-10528)

Longer term for create_all_vm, it may be better to eliiminate it,
relying instead on the nearly-identical pool/create_all_hw.py tests.

Signed-off-by: Kenneth Cain <kenneth.c.cain@intel.com>