DAOS-10770 test: deployment/agent_failure.py - Update IOR error expectations - release/2.2 (#9273)
test_agent_failure is failing because -1005 isn't in the IOR results.
The error message is different from when I created this test.

test_agent_failure_isolation is failing because IOR on the kill
client is failing. When I created this test, no error was observed
on the client where the agent was killed.

The resolution is to fix the test so that it expects no
error or different error message.

Skip-unit-tests: true
Skip-fault-injection-test: true
Test-tag: agent_failure
Signed-off-by: Makito Kano <makito.kano@intel.com>