CART-89 multiprovider: Add new api to query original src provider
Add new api to query whether the originators provider was primary
or secondary.
For server originators, this will depend on the context on which rpc
was allocated as server can support both a primary and a secondary
context.
For client originators, this will depend on CRT_SECONDARY_PROVIDER
envariable. If set it will indicate that the client cluster is
using a secondary provider.

Signed-off-by: Alexander A Oganezov <alexander.a.oganezov@intel.com>