DAOS-9784 dfuse: Allow True/False for dfuse cache settings.
Allow True/False as well as on/off and ignore case for dfuse
cache attributes.
Ensure all attribute settings are logged.
Correct an error when an invalid setting was used for direct-io-disable.

This makes it more user friendly, as well as allowing Python True/False
values to be set without translating into on/off.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>