DAOS-10601 client: fix daos_event_abort corruption issue (#9175)
daos_event_abort() was marking the event as aborted, however the DAOS
internal task is not aborted, and currently there is no way to abort
that. So later if a user progresses (through test or poll) that aborted
event, the event API frees the event and marks it as ready to initialized
since it was aborted before, but the DAOS internal task is still not done.

When the DAOS task finally progresses, it can cause corruption because the
user might have resused the event, or even reused the buffers that were passed
to the event.

Since we do not support cancellations of internal DAOS tasks and RPCs, change
the abort implementation to just be a no-op for now which would cause a test or poll
on the event to wait till the internal task is completed to avoid such corruption issues.

Signed-off-by: Mohamad Chaarawi <mohamad.chaarawi@intel.com>