DAOS-623 build: Improve the preprocessing steps (#9390)
A couple of things were causing the preprocessing step in cart
to generate _structures_from_macros.h to be slow

1. The Preprocessor builder currently doesn't support building
   multiple items in parallel so it's better to call it one file
   at a time.
2. Use Command builders for intermediate steps so they don't need
   to be repeated
3. Pick to allow W503

Also cherry-picks this patch:
DAOS-623 build: Update clang-format settings. (#9007)

Change Alignment to be closer to our established code.
Change spaces to tabs to avoid a clang bug.
Enable clang-format githook if clang-format 14.0.5 or greater is detected

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>