DAOS-10870 dtx: handle active DTX entry update properly
For any transaction, the DTX entry in the active DTX table is unique.
So when we allocate new DTX entry in the active DTX table, we should
not hit the entry with the same DTX ID. If that happened, then there
may be two possible cases:

1. Client resent the RPC but without set 'RESEND' flag.
2. Client reused the DTX ID for different modifications.

Currently, the 1st case is more suspected which is DAOS client logic
bug, need to be fixed correspondingly. If it is the 2nd case, we can
re-generate new TX ID.

Signed-off-by: Fan Yong <fan.yong@intel.com>