CART-89 multiprov: Auto-tag replacement for secondary providers
- When the client is secondary, we do not want to use target endpoint
information for deciding which secondary tag to send rpc to.
Instead we send it to secondary tag=0. In future API will be added
to control total number of server tags, and instead of 0 a random
tag will be chosen.

intended destination tag is present in the rpc header and available
to be retrieved via apis.

- test updated to try new apis and test behavior

Signed-off-by: Alexander A Oganezov <alexander.a.oganezov@intel.com>