DAOS-10610 dfuse: Accept positional args.
Add support for positional args to dfuse for mountpoint, pool,
container.

If dfuse is going to be mounted via fstab then exactly one
option is given which is mountpoint so put this first, but also
accept pool and container if given.

This allows dfuse to be mounted via the standard mount command and
fstab via a line like the following:

dfuse /mnt/dfuse fuse3 defaults,user 0 0

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>