DAOS-7821 gha: Add code to report Jira status into GitHub. (#9628)
Add a GitHub action to parse Jira ticket information for a PR and update the PR
metadata to keep it in sync with Jira. Add a comment to the PR which will be edited
to keep it up-to-date, and report the ticket description and status as well as adding
a link.

Warn on several issues around malformed PR titles (which become part of the commit
message when landed) and if work is happening on tickets which are not marked
as In Progress or In Review.

Set labels for 2.2 and 2.4 release tickets, as well as anything that should be treated
as a priority.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>