DAOS-10929 container: fix size of cont_global_version value (#9545)
Before this change, when a new container was created, the metadata
rdb value for key ds_cont_prop_cont_global_version was written as
a 64-bit value in cont_prop_write(). However, the intent for this
recently-added key is a uint32_t typed value. Also, an inconsistency
exists with the pool/container layout upgrade code, upgrade_cont_cb()
writing the value as uint32_t. This results in some containers with
64-bit version values, and others with 32-bit values.

With this change, the cont_prop_write() and cont_prop_read() functions
are changed to write and read a 32-bit cont_global_version value in rdb,
to agree with the upgrade_cont_cb() code.

Signed-off-by: Kenneth Cain <kenneth.c.cain@intel.com>