DAOS-11031 control: Clean hugepages on engine startup (#9625)
The previous approach to cleanup hugepages created by SPDK running in
engine process was to remove in a callback triggered on engine exit.
The criteria for hugepage file removal was that file owner matched the
running user UID and the PID encoded in the filename matched that of the
exited engine process.

This approach will not always remove all hugepages created by DAOS
processes, for example in the case where daos_server is sent SIGKILL or
when different users execute daos_server.

A more aggressive approach is implemented in this PR whereby, on
completion of engine storage format, a callback is triggered which removes
all hugepages created by SPDK that belong to a no-longer-active process
(identified by PID encoded in the filename).

Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>