DAOS-11113 test: Fix config_gereate_run.py by removing warning message in generated config
The test uses dmg config generate to generate a new server
config and tries to start server with it. However, the
generated config contains a warning message in the first
line when there's only one AP. If we include this warning
line, the server would fail to start. Fix the test script
so that it removes this line and tries to start server
with the rest.

Skip-unit-tests: true
Skip-fault-injection-test: true
Test-tag: config_generate_run
Signed-off-by: Makito Kano <makito.kano@intel.com>