SRS-371 csum: Checksum Scrubbing Feature (#9345)
This is a single commit for the checksum scrubbing feature that was
developed on the feature/csum-scrubbing feature branch. It includes
the implementation and functional testing. The following tickets were
completed for this feature:

DAOS-5465: Corrupt BIO Address through VOS
Add the ability to set a bio address as corrupted using the VOS
iter API. This will be necessary for the checksum scrubber
to be able to mark data as corrupted when silent data
corruption is detected.

DAOS-6071: Introduced new pool and container properties as defined in
data_integrity.md

DAOS-6049: Checksum Scrubber Implementation: It iterates the vos
objects and for each data value, it will calculate a checksum and
compare it to the stored checksum. Added CORRUPTION_DETECTED to the
RAS_EVENT_LIST in preparation for raising RAS events. Introduced new
telemetry metrics. The scrubber is split between two files:
srv_pool_scrub_ult.c and vos_pool_scrub.c. The "ult" file is called
by the server to start the scrubber in its own persistent background
ult. The vos file does the actual scanning and scrubbing of the VOS
tree. Added a new test executable (srv_scrubbing_tests) which includes
unit tests for the vos scrubbing functionality.

DAOS-5467: Auto Evict Pool Target - When the checksum scrubber
detects enough silent data corruptions to exceed a threshold it will
initiate a drain on the pool target. Updated the migration system so
that if it detects a checksum error it will not fail the
migration/rebuild but will simply skip that object.

Testing introduced:

DAOS-8099: Enabling the scrubber in some basic tests and OSA tests

DAOS-9318: Target eviction test case

DAOS-8417 New functional tests for scrubber that read the scrubber
telemetry metrics and uses fault injection to verify that scrubbing
can be enabled and that data corruption is detected.

DAOS-7300: Target auto eviction

Signed-off-by: rpadma2 <ravindran.padmanabhan@intel.com>
Signed-off-by: Ryon Jensen <ryon.jensen@intel.com>
Co-authored-by: Jeff Olivier <jeffrey.v.olivier@intel.com>