DAOS-11155 test: Increase pool size for container/multiple_delete.py
Pool size was too small; 44GB/rank with 8 targets.
We write 2GB of data with IOR and that could get
out of space, so increse the size to 10%, which
is about 320GB(SCM) + 320GB(NVMe)/rank in this
configuration.

Don't create a new container during IOR run because
we have already created before running it.

Refactor the code.

Skip-unit-tests: true
Skip-fault-injection-test: true
Test-tag: multi_container_delete agent_failure_basic
Signed-off-by: Makito Kano <makito.kano@intel.com>