DAOS-11077 dfs: fix mtime accounting for user set mtime (#9636)
In some data copying tools, the file is updated and the mtime is reset on the file after the update is complete. This patch account for a user set mtime and returns the latter of either the mtime update or the file modfication.
Using HLC converted from timespec allows for nsec precision.

Signed-off-by: Mohamad Chaarawi <mohamad.chaarawi@intel.com>
Co-authored-by: Ashley Pittman <ashley.m.pittman@intel.com>