DAOS-11073 control: Don't remove pool label during destroy (#9746)
As part of the work done for DAOS-8237, the pool destroy RPC
handler removed the pool service label if the request was
forced. This seemed like a usability enhancement, in that
a forced destroy would always succeed from the perspective
of the admin, particularly in the case of wanting to reuse
the label. In cases where the pool was left in a destroying
state, the idea was that it would eventually be cleaned up
as part of pool creation or leadership step-up.

In hindsight, that change was probably a bad idea in that it
hides problems with the pool destroy flow at the engine level
and could lead to resource leaks. Better to leave the label
intact on the pool service so that it's clear to the admin
that the pool isn't really destroyed yet. The next challenge
is in figuring out how to convey to the admin why the pool
can't be destroyed, even if the force flag has been set.

Signed-off-by: Michael MacDonald <mjmac.macdonald@intel.com>