DAOS-11134 control: Add leadership check to doGroupUpdate() (#9737)
In rare circumstances, this method could be called after
leadership is lost but before the worker loop is exited.
Add a final check for leadership before invoking the group
update dRPC in order to avoid trying to make the group
update on a non-leader replica.

Also fixes a bug where a sync group update could be followed
by an unnecessary async group update.

Signed-off-by: Michael MacDonald <mjmac.macdonald@intel.com>