DAOS-11180 dtx: re-index committed DTX entries for DTX_CHECK
If the DTX_CHECK RPC handler hits -DER_INPROGRESS failure on some
target, then it means that the committed DTX entries on such target
are not re-indexed. Under such case, we need to trigger DTX re-index
on the target for subsequent (retry) DTX_CHECK RPC. Otherwise, it may
cause more DTX_CHECK RPCs to be retried (for ever), as to related DTX
resync cannot complete on some DTX leaders.

Signed-off-by: Fan Yong <fan.yong@intel.com>