DAOS-10138 pool: Try to restore PS redundancy (#9119)
Currently, the pool service (PS) only replaces DOWN replicas. If there
are not enough engines in the pool, the PS runs at a degraded
redundancy. When the number of available engines in the pool increases,
the PS does not, however, attempt to restore its redundancy. This patch
slightly improves this policy to try to restore some redundancy toward
a new PS RF property, svc_rf.

For new pools, this new PS RF property defaults to 2; for existing
pools, this new property defaults to 2 or current effective RF (e.g., 1
for 3 or 4 replicas), whichever is greater. It can be specified at pool
creation time only; we'll support changing it later.

We limit the scope of this patch to the restoration policy, not the
mechansim.

Signed-off-by: Li Wei <wei.g.li@intel.com>