DAOS-11266 test: Implement TargetedExecutableCommand to run ddb on server host
ddb needs to run on a server host, but current implementation
of ExecutableCommand runs the command on the test host, so
create a new class, TargetedExecutableCommand, to run the
command on a specified host. In the new class, use run_pcmd()
as it takes host as parameter.

After running the command, construct CmdResult from the output
to be consistent with ExecutableCommand.run().

Created a new class rather than extending ExecutableCommand
because ExecutableCommand has several obsolete and
unnecessary features.

Add server_host parameter to DdbCommand and DdbCommandBase
class.

Skip-unit-tests: true
Skip-fault-injection-test: true
Test-tag: pr

Signed-off-by: Makito Kano <makito.kano@intel.com>