DAOS-11231 dtx: handle dependency during task conversion
For the task that need to be handled via distributed transaction,
the client side IO logic will convert it to a transactional task.
During dc_tx_convert(), the original task maybe freed because of
the dependency on the sub-task(s) for existence check inside the
dc_tx_attach(). That will cause the "task" to become invalid for
subsequent process for the conversion.

To avoid above bad case, the patch moves the TX commit logic for
task conversion into related existence check callback before the
TX release.

Signed-off-by: Fan Yong <fan.yong@intel.com>