DAOS-10803 test: Update the object class and RF. (#7690) (#9341)
Backport of the patch 747fd6d updating the object class and RF for
offline_reintegration.

The oclass RP_2G8 is not adapted for the test
osa/offline_reintegration.py:offline_reintegration_oclass.  Indeed, some
shards of the ior test file could be lost during the process of the many
cycles of rebuild/reintegration. These lost shards, could eventually
makes the fds_open call of ior to fail.

More details on this issue are available on the DAOS-9313 ticket.

Signed-off-by: Cedric Koch-Hofer <cedric.koch-hofer@intel.com>
Co-authored-by: Ravindran Padmanabhan <ravindran.padmanabhan@intel.com>