DAOS-9602 chk: verify container label
If the container label in the container service (cont_svc::cs_uuids)
exists but does not match the label in the container property, then
trust the container service and reset the one in container property.

Otherwise, if the container label in the container service does not
exists, but the one in the container property is there, then trust
the label in the container property and add it to container service.

Signed-off-by: Fan Yong <fan.yong@intel.com>