DAOS-11035 dtx: fix some DXT related issues
It includes the following fixes:

1. Skip new non-committed DTX when check visibility for rebuild

Up layer rebuild logic guarantees that the rebuild scan will not be
triggered until DTX resync has been done on all related targets. So
here, if rebuild logic hits non-committed DTX entry, it must be for
new IO that version is not older than rebuild, then it is invisible
to rebuild. Related new IO corresponding to such non-committed DTX
has already been sent to the in-rebuilding target.

2. Skip new DTX during DTX resync

Even if current target was the DTX leader, but if the version of
the DTX is older than DTX resync version, we still need to resync
it. Because its sponsor may be dead (if not dead, then skip it).

3. Pin DTX entry in DRAM when DTX start

Originally, we pin the DTX entry in DRAM conditionally before ULT
yield. But it is relative complex to forecast whether the ULT may
yield or not. This patch much simplifies related logic and avoids
missing some corner cases.

4. Validate DTX entry before become committable or being committed

If the DTX entry is returned via the iteration. The upper layer caller,
for DTX resync or DTX cleanup, may abort it sometime later. Before being
aborted, it may re-attach to some DTX handle for handling resent request.
Then there is race between them. So before become committable or being
committed by RPC handler, needs to check whether it has been aborted or
not.

5. Unify DTX resync for pool map refresh and container open

That will simplify DTX resync logic, cleanup confused parameter
"resync_all", and avoid missing to handle some DTX entry if need
DTX resync for both cases by race.

Signed-off-by: Fan Yong <fan.yong@intel.com>