DAOS-11281 tests: Update pool/svc.py expectations
Pull request #9119 has changed the default number of PS replicas from 3
to 5 via the new pool property svc_rf. It has also disallowed the
specification of even numbers of PS replicas. This patch updates the
expectations of the pool/svc.py tests to reflect those changes.

The old rule that avoids rank 0 when selecting PS ranks causes the
maximum number of PS ranks to be 4, instead of 5, for 5-rank pools that
include rank 0. The plan is to remove that rule via the deployment of
multiple MS replicas in testing. Hence, we simply expect 4 PS ranks for
svc_params_none, svc_params_0, and svc_params_5 in this patch.

Test-tag: pool_svc
Signed-off-by: Li Wei <wei.g.li@intel.com>
Required-githooks: true