DAOS-10138 pool: Improve PS reconfigurations
A PS currently performs reconfigurations (i.e., membership changes) only
upon pool map changes. If the PS leader crashes in the middle of a
series of reconfigurations, the new PS leader will not plan any
reconfiguration (or notify the MS of the latest list of replicas) until
the pool map changes for some other reason. This patch lets a PS leader
check if reconfigurations are required when it steps up.

To avoid blocking the step up and pool map change processes, this patch
performs each series of reconfigurations asynchronously in a ULT. The
change allows the reconfiguration process to wait for pending events,
retry upon certain errors (in the future), and wait for RPC timeouts
without directly impacting the normal PS operations. Hence, this patch
reverts the workaround (209ba92bf07e84d80ded669dd3c67683e47f9772) that
skips destroying PS replicas.

Moreover, this patch adds a safety net that prevent an older rsvc leader
from removing an rsvc replica created by a newer rsvc leader. Although
it cannot resolve all problems in the area, the natural, term-based
approach requires no RDB layout change and is simple to implement. The
patch has to change rdb_test a bit to allow more than one test rsvc
instance, so that a quick rsvc test can be added.

Since select_svc_ranks avoids rank 0 but ds_pool_plan_svc_reconfs does
not, this patch modifies the former to remove the avoidance, so that
during a PoolCreate operation the MS observes a notification from the
new PS with the same ranks as the PoolCreate dRPC response. We have to
change a few tests as well as the MS to make this work:

  - Work around a race between mgmtSvc.PoolCreate and
    Database.handlePoolRepsUpdate. See the comment for the details.

  - Update svc.yaml to reflect the new PS replacement policy.

  - Fix the daos_obj.c assertion that PS leaders can be rank 0.

Signed-off-by: Li Wei <wei.g.li@intel.com>
Required-githooks: true