DAOS-11186 tests: Bump crt_timeout for boundary.py
The 10-s crt_timeout seems to be too short for boundary.py, which works
with 20,000 or 30,000 containers. See the Jira ticket for the numbers.
This patch bumps the crt_timeout used by boundary.py to 60 s.

In addition, this test requires a larger DAOS_MD_CAP value that seems to
be provided by some special weekly settings. In order to make the test
work in non-weekly settings, this patch adds the DAOS_MD_CAP value
explicitly.

Test-tag: container_boundary
Signed-off-by: Li Wei <wei.g.li@intel.com>
Required-githooks: true