DAOS-11266 test: Extend CommandWithParameters for DdbCommandBase (#9906)
ddb needs to run on a server host, but current implementation
of ExecutableCommand runs the command on the test host, so
create a new class, TargetedExecutableCommand, to run the
command on a specified host. In the new class, use run_pcmd()
as it takes host as parameter.

After running the command, construct CmdResult from the output
to be consistent with ExecutableCommand.run().

Created a new class rather than extending ExecutableCommand
because ExecutableCommand has several obsolete and
unnecessary features.

Add server_host parameter to DdbCommand and DdbCommandBase
class.

Extend more primitive ExecutableCommand and implement
run() that simply runs run_pcmd(). Now the return type
is the same as that of run_pcmd(), so update the comment
in ddb_utils.py

Also consolidate ddb_utils.py and ddb_utils_base.py

Signed-off-by: Makito Kano <makito.kano@intel.com>