DAOS-9585 chk: Fix checker enable/disable workflows (#10047)
Make system checker state persistent by storing it in
the system database so that the system can be stopped
and restarted while staying in checker mode.

Break out check enable/disable into discrete operations,
and rework rank join logic to receive checker state in
the join response. If the rank was not started in checker
mode (e.g. because the local node was restarted), restart
the rank after enabling checker mode.

Signed-off-by: Michael MacDonald <mjmac.macdonald@intel.com>