DAOS-11035 dtx: fix some DXT related issues (#10056)
It includes the following fixes:

1. Skip new non-committed DTX when check visibility for rebuild

Up layer rebuild logic guarantees that the rebuild scan will not be
triggered until DTX resync has been done on all related targets. So
here, if rebuild logic hits non-committed DTX entry, it must be for
new IO that version is not older than rebuild, then it is invisible
to rebuild. Related new IO corresponding to such non-committed DTX
has already been sent to the in-rebuilding target.

2. Skip new DTX during DTX resync

Even if current target was the DTX leader, but if the version of
the DTX is older than DTX resync version, we still need to resync
it. Because its sponsor may be dead (if not dead, then skip it).

Signed-off-by: Fan Yong <fan.yong@intel.com>