DAOS-10301 test: Fix daos_test fork interactions
The primary fix is the usage of the environment variable IBV_FORK_SAFE
in daos_test runs kicked off by ftest. This is needed due to a known
issue with libibverbs, wherein a forked process may accidentally touch
a memory page registered for communications. This causes a segfault.
See: https://www.open-mpi.org/faq/?category=openfabrics#ofa-fork

This patch also includes additional miscellaneous fixes that are
not causes of this issue, but they improve debugability and
correctness, in the event that we see any further issues in this
area.

- Set IBV_FORK_SAFE environment variable for daos_test runs.
- In dmg test helper for daos_test, explicitly check EOF instead
  of assuming 0 bytes means we're done.
- Fixed a memory leak in pool tests.
- Use execvp() instead of system().
- Fix potential segfault while reading stderr pipe.
- Assume failure if exec returns.
- Add a barrier to each call of the subtest.

Test-tag: test_daos_pool
Test-repeat: 50
Quick-functional: true

Signed-off-by: Kris Jacque <kristin.jacque@intel.com>