DAOS-11472 control: Compare nr backing devs across engines if VMD enabled
At server startup, it is checked (and enforced) that the number of
entries in the bdev_list is identical for all engines. In VMD configs
this mechanism is comparing the number of VMD endpoint addresses, not
NVMe backing devices.

In a scenario where the server has 3 VMD domains: one with 4 NVMe disks
and two with 2 NVMe disks each and the engine configs have 1 and 2 VMD
endpoint addresses each, the number of NVMe backing devices is equal.

In such a scenario the validation should not fail so rather than
comparing the number of raw bdev_list entries, instead compare the total
number of NVMe backing devices per engine if VMD is enabled.

Features: control
Required-githooks: true

Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>