DAOS-11476 test: Fix coverity unchecked return in jobtest
Before this change the jobtest invocation of dc_cont_hdl2uuid() did
not check for a failure, leading to Coverity CID 22078. With this
change, the return code is checked and the test output is adjusted.

Performing build only in CI (skipping testing) since jobtest is only
invoked manually at this time. It was tested manually with this change.

Skip-test: true

Required-githooks: true

Signed-off-by: Kenneth Cain <kenneth.c.cain@intel.com>