DAOS-11477 pool: Coverity CIDs 105913/105918 always free prop
Proactive cherry-pick of same change in progress for master branch.

In pool_upgrade_props() and ds_pool_mark_upgrade_completed() it is
possible that the pool_prop_read() call allocates memory for the
daos_prop_t * output argument (even when rc != 0 is returned).
Before this change, prop was not freed, and with this change
it is freed even when pool_prop_read() fails.

Features: pool

Required-githooks: true

Signed-off-by: Kenneth Cain <kenneth.c.cain@intel.com>