DAOS-7332 control: allow hostname with negative port
Adjusted the parse function in hostlist.go to allow hostnames that are
provided with a negative port number. A hostname with a negative port
number is now accepted into the hostlist instead of throwing
an "invalid hostname" error. The "negative port" unit test in
hostlist_test.go was modified to expect this new behavior.

Signed-off-by: Margaret Lawson <mlawsonca@google.com>