DAOS-11232 vos: Remove unecessary and incorrect assertion (#10232)
The assertion in question is incorrect.  A set of valid inputs
can be chosen to make it fail.  I'm in process of creating a
regression test and checking in my utility that I wrote while
debugging this but it's taking too long so let's fix the issue
only first.

Required-githooks: true

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>