DAOS-9585 chk: Reset findings DB on checker start
When the checker is starting a new pass, reset the findings
DB so that new findings are more apparent. Another option, if
feedback guides it, is to keep older findings but by default
only show findings since the most recent start. The tools
could optionally show all findings if desired.

Required-githooks: true

Signed-off-by: Michael MacDonald <mjmac.macdonald@intel.com>