DAOS-11363 test: Apply threshold for SCM free space check in test (#9974)
Increase timeout for check_free_space for SCM at the end of 
container/multiple_delete.py

After 50 iterations of container create, IOR run, container destroy,
SCM free space isn’t fully recovered. However, this may be due to
the aggregation taking a while. If we wait for a while before the
final free space check, the space may be fully recovered.

Increase the timeout for check_free_space() to 180 sec, pass in the
SCM initial value, and see if the test passes. (Default timeout is 30 sec)

Signed-off-by: Makito Kano <makito.kano@intel.com>