DAOS-10365 dfuse: Allow dfuse caching of metadata. (#9565)
Stash the stat information for inodes and if the kernel re-requests the data within
the validity timeout then reply with the existing data and a timeout value to match
the original expiry date.
This is mostly for when writeback caching is enabled where the kernel will make
requests for metadata updates despite being given permission to cache the data.
Only use this cache when the file hasn't been written to since last update, and
disable when interception library is in use.

Use cached values for inodes when required.
When the kernel is using writeback cache then it can query inode
metadata before the timeout has expired and this costs every time
a file is opened.  To handle this keep a record of when the inode
was last reported to the kernel and for requests before the
timeout then reply with the cached data, but use a shorter timeout.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>